requirements_minimal.txt
# Minimal requirements without voice support
discord.py==2.3.2
python-dotenv==1.0.0
spotipy==2.23.0 